# Compilation Speedup (~2.3 seconds) - Reduces compilation overhead by disabling icon-cutter & DMTarget inputs/outputs