<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull request process. -->

## About The Pull Request

- Removes icon-cutter from the compilation step since we don't actually use it and it introduces A LOT of overhead.
- Changes it so that DM will always compile. GLOBing every single code file in the codebase adds a small amount of overhead.
- Executes DM step before TGUI step, which will probably save a zero amount of time.

People with faster CPUs (and ones with turboboost) will see less improvements than I am seeing since I have quite an old PC.

## Why It's Good For The Game

Our compilation time is starting to get a bit ridiculous and there are some easy savings to be had by removing high-overhead systems from the build steps.

## Testing Photographs and Procedure

All tests were done with the same programs opened and visual studio focused. (Unfocusing it increases compilation times by ~2x).

Overhead was measured by taking the total build time and subtracting the DM target build time from it.

### Baseline

|Test-Run|Overhead (/s)|
|-|-|
|1|2.632|
|2|2.543|
|3|2.634|

### With Icon-Cutter Change Only

|Test-Run|Overhead (/s)|
|-|-|
|1|2.208|
|2|2.174|

~0.4 seconds saved

### With DM-Step Change Only

|Test-Run|Overhead (/s)|
|-|-|
|1|1.299|
|2|1.323|
|3|1.296|

~1.3s saved

### After

|Test-Run|Overhead (/s)|
|-|-|
|1|0.337|
|2|0.352|
|3|0.344|

~2.3s saved

If you do the math there are 0.6s of savings unaccounted for, so I will chalk that down to re-ordering the DMtarget and the TGUI target.

## Changelog
:cl:
tweak: Improves compilation times by ~2.3 seconds by removing unnecessary build process overhead.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put your name to the right of the first :cl: if you want to overwrite your GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the icon ingame) and delete the unneeded ones. Despite some of the tags, changelogs should generally represent how a player might be affected by the changes rather than a summary of the PR's contents. -->