@PowerfulBacon 2024-09-28
Actually lets be honest, we are never going to actually use icon cutter so I'm just deleting it rather than commenting it out.