@PowerfulBacon 2024-10-04
The compilation time is starting to annoy me